+++ client/Packages/com.beamable.server/Editor/UI/Components/ServiceBaseVisualElement/ServiceBaseVisualElement.cs
@@ -277,7 +277,7 @@ namespace Beamable.Editor.Microservice.UI.Components
 		protected virtual void SetupProgressBarForStart(Task task)
 		{
 			// We have two ways. Either store reference or return instance as event parameter
-			new RunImageLogParser(_loadingBar, Model);
+			new RunImageLogParser(_loadingBar, Model.Builder, Model.Name);
 		}
 		private void OnStartFailed()
 		{