+++ client/Packages/com.beamable.server/Editor/UI/Model/MicroservicesDataModel.cs
@@ -5,7 +5,6 @@ using Beamable.Server.Editor.UI.Components;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using UnityEngine;
 
 namespace Beamable.Editor.UI.Model
 {