+++ client/Packages/com.beamable.server/Editor/UI2/UsamWindow.cs
@@ -1,19 +1,25 @@
-using Beamable;
 using Beamable.Common;
+using Beamable.Editor.Microservice.UI;
 using Beamable.Editor.Microservice.UI.Components;
 using Beamable.Editor.Microservice.UI2.Components;
+using Beamable.Editor.Microservice.UI2.Models;
+using Beamable.Editor.Toolbox.Components;
+using Beamable.Editor.Toolbox.Models;
 using Beamable.Editor.UI;
+using Beamable.Editor.UI.Model;
 using Beamable.Server.Editor;
 using Beamable.Server.Editor.Usam;
+using System.Runtime.Remoting.Contexts;
 using UnityEditor;
+using UnityEngine;
 using UnityEngine.UIElements;
-using Usam;
 
 namespace Beamable.Editor.Microservice.UI2
 {
 	public class UsamWindow : BeamEditorWindow<UsamWindow>
 	{
 		private CodeService _codeService;
+		private UsamDataModel _dataModel;
 
 		static UsamWindow()
 		{
@@ -40,7 +46,9 @@ namespace Beamable.Editor.Microservice.UI2
 		private VisualElement _windowRoot;
 		private ActionBarVisualElement _actionBarVisualElement;
 		private MicroserviceBreadcrumbsVisualElement _microserviceBreadcrumbsVisualElement;
+		private CreateServiceVisualElement _createServiceElement;
 
+		private ScrollView _scrollView;
 
 		protected override void Build()
 		{
@@ -60,31 +68,74 @@ namespace Beamable.Editor.Microservice.UI2
 
 			_actionBarVisualElement = root.Q<ActionBarVisualElement>("actionBarVisualElement");
 			_actionBarVisualElement.Refresh();
+			_actionBarVisualElement.UpdateButtonsState(_codeService.ServiceDefinitions.Count);
+
 			_actionBarVisualElement.OnRefreshButtonClicked += HandleRefreshButtonClicked;
+
+
 
 			_microserviceBreadcrumbsVisualElement = root.Q<MicroserviceBreadcrumbsVisualElement>("microserviceBreadcrumbsVisualElement");
 			_microserviceBreadcrumbsVisualElement.Refresh();
-			var scrollView = new ScrollView(ScrollViewMode.Vertical);
+			_scrollView = new ScrollView(ScrollViewMode.Vertical);
 			var emptyContainer = new VisualElement { name = "listRoot" };
 
 			var microserviceContentVisualElement = root.Q("microserviceContentVisualElement");
-			microserviceContentVisualElement.Add(scrollView);
-			scrollView.Add(emptyContainer);
+			microserviceContentVisualElement.Add(new VisualElement { name = "announcementList" });
+
+			_createServiceElement = new CreateServiceVisualElement();
+			_createServiceElement.SetHidden(true);
+			microserviceContentVisualElement.Add(_createServiceElement);
+
+			microserviceContentVisualElement.Add(_scrollView);
+			_scrollView.Add(emptyContainer);
 			OnLoad().Then(_ =>
 			{
+				_dataModel = Scope.GetService<UsamDataModel>();
+				if (!_codeService.IsDockerRunning)
+				{
+					ShowDockerNotRunningAnnouncement();
+					return;
+				}
+
 				foreach (BeamoServiceDefinition beamoServiceDefinition in _codeService.ServiceDefinitions)
 				{
+					var model = _dataModel.GetModel(beamoServiceDefinition.BeamoId);
 					var el = new StandaloneMicroserviceVisualElement() { Model = beamoServiceDefinition };
 					emptyContainer.Add(el);
+
+					model.OnLogsDetached += () => { ServiceLogWindow.ShowService(model); };
+					if (!model.AreLogsAttached)
+					{
+						ServiceLogWindow.ShowService(model);
+					}
 					el.Refresh();
 				}
 			});
+
+			_actionBarVisualElement.OnCreateNewClicked += HandleCreateNewButtonClicked;
+		}
+
+		private void ShowDockerNotRunningAnnouncement()
+		{
+			var dockerAnnouncement = new DockerAnnouncementModel() { IsDockerInstalled = true, OnInstall = Build };
+			var announcementList = _windowRoot.Q<VisualElement>("announcementList");
+			announcementList.Clear();
+
+			var element = new DockerAnnouncementVisualElement() { DockerAnnouncementModel = dockerAnnouncement };
+			announcementList.Add(element);
+			element.Refresh();
 		}
 
 		private void HandleRefreshButtonClicked()
 		{
 			_codeService = Scope.GetService<CodeService>();
-			_codeService.RefreshServices().Then(_ => { });
+			_codeService.RefreshServices().Then(_ => Build());
+		}
+
+		private void HandleCreateNewButtonClicked(ServiceType serviceType)
+		{
+			_createServiceElement.Refresh(_actionBarVisualElement.Refresh);
+			EditorApplication.delayCall += () => _scrollView.verticalScroller.value = 0f;
 		}
 
 		public override async Promise OnLoad()