+++ client/Packages/com.beamable.server/Editor/Usam/Editor/RegisterDependencies.cs
@@ -1,5 +1,4 @@
 using Beamable.Common.Dependencies;
-using Beamable.Editor;
 
 namespace Beamable.Server.Editor.Usam
 {