+++ client/Packages/com.beamable/Editor/BeamEditor.cs
@@ -232,7 +232,15 @@ namespace Beamable
 			// Reload the current environment data
 			BeamableEnvironment.ReloadEnvironment();
 
-			BeamCliUtil.InitializeBeamCli();
+			try
+			{
+				BeamCliUtil.InitializeBeamCli();
+			}
+			catch
+			{
+				EditorApplication.delayCall += Initialize;
+				return;
+			}
 
 			// If we ever get to this point, we are guaranteed to run the initialization until the end so we...
 			// Initialize Editor instances of Reflection service